Add WebController
We want to add a new FormDocumentsController which will be an API
endpoint called by forms-runner returning JSON, rather than a webpage
as all our other controllers return.

We don't want to inherit from our current ApplicationController, as
this has callbacks related to user authentication and requests made to
get a webpage.

Move all the behaviour of ApplicationController that is not relevant
to API endpoints into a new WebController, and make all existing
controllers inherit from that.

Keep the callback to set up the common logging attributes in
ApplicationController so our new FormDocumentsController can inherit
from this.
