+++ app/controllers/errors_controller.rb
@@ -1,4 +1,4 @@
-class ErrorsController < ApplicationController
+class ErrorsController < WebController
   skip_before_action :authenticate_and_check_access, except: :forbidden
   skip_before_action :check_maintenance_mode_is_enabled, only: :maintenance
   skip_before_action :redirect_if_account_not_completed
