+++ app/controllers/forms/contact_details_controller.rb
@@ -1,5 +1,5 @@
 module Forms
-  class ContactDetailsController < ApplicationController
+  class ContactDetailsController < WebController
     after_action :verify_authorized
     def new
       authorize current_form, :can_view_form?
