+++ app/controllers/reports_controller.rb
@@ -1,4 +1,4 @@
-class ReportsController < ApplicationController
+class ReportsController < WebController
   before_action :check_user_has_permission
   after_action :verify_authorized
