+++ app/controllers/users_controller.rb
@@ -1,4 +1,4 @@
-class UsersController < ApplicationController
+class UsersController < WebController
   after_action :verify_authorized
   after_action :verify_policy_scoped, only: :index
